MyPlayground, Xcworkspace, and a Video added
Making use of UILabel animations and a set of UIColor that WWDC landing
page use, this is a simple app giving an overview of WWDC17 as the user
taps on each UILabel. The UILabel animation used is an open source
project at GitHub, named LTMorphingLabel. Along with Swift playground,
I’ve also attached the project files and a screencast of the app as
assets.